Moved common shader interop structures to dedicated header
